CL0004

+++ store/src/main/java/com/michaelspringproject/store/dtos/ProductDto.java
@@ -9,5 +9,7 @@ public class ProductDto {
     private String productName;
     private String description;
     private double price;
-    private int categoryId;
+    private Long categoryId; 
+    
 }
+
